html/semantics/forms/the-form-element/form-autocomplete.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>form autocomplete attribute</title>
<link rel="author" title="Denis Ah-Kang" href="mailto:denis@w3.org">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#the-form-element">
<link rel=help href="http://www.w3.org/html/wg/drafts/html/master/#attr-fe-autocomplete">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<form name="missing_attribute">
  <input>
  <input autocomplete="on">
  <input autocomplete="off">
</form>
<form name="autocomplete_on" autocomplete="on">
  <input>
  <input autocomplete="on">
  <input autocomplete="off">
</form>
<form name="autocomplete_off" autocomplete="off">
  <input>
  <input autocomplete="on">
  <input autocomplete="off">
</form>
<form name="autocomplete_invalid" autocomplete="foobar">
  <input>
  <input autocomplete="on">
  <input autocomplete="off">
</form>
<script>
  function autocompletetest(form, expectedValues, desc) {
    test(function(){
      assert_equals(form.autocomplete, expectedValues[0]);
      assert_equals(form.elements[0].autocomplete, expectedValues[1]);
      assert_equals(form.elements[1].autocomplete, expectedValues[2]);
      assert_equals(form.elements[2].autocomplete, expectedValues[3]);
    }, desc);
  }

  autocompletetest(document.forms.missing_attribute, ["on", "on", "on", "off"], "form autocomplete attribute missing");
  autocompletetest(document.forms.autocomplete_on, ["on", "on", "on", "off"], "form autocomplete attribute on");
  autocompletetest(document.forms.autocomplete_off, ["off", "off", "on", "off"], "form autocomplete attribute off");
  autocompletetest(document.forms.autocomplete_invalid, ["", "", "on", "off"], "form autocomplete attribute invalid");
</script>
